#26 3D Game Engine Tutorial: Specular Reflection
Implementation of the "Specular Reflection" Tutorial:
https://www.youtube.com/watch?v=ZYb0Lg6ko9k

Note:
The Normals should be a property of the face (vectors indexes) and not
the vertex itself (in a pyramid each vertex should have 3 normals, one
for each face).
This cause an issue with the Specular reflection in this tutorial (more
evident if you face the reflective vector).
Anyway, I'm trying to be as close as possible to this great tutorial.