#28 3D Game Engine Tutorial: Point Lights
Implementation of the "Point Lights" Tutorial:
Part1:
https://www.youtube.com/watch?v=t47DohOx6OI
Part2:
https://www.youtube.com/watch?v=EEDYe6rChq8

Note:
Due to an issue I'm having using an array of structs as uniform in the
fragment shader, in this commit I resolved the array ( pointLights[4] )
in 4 different variables ( pointLights0 ... pointLights3 ). This solve
the issue, I hope to find a solution or an explaination soon.